I added files: licenses, requirements.txt and README.md.
In the license file, I specified the MIT license.
In requirements.txt, I specified the Pillow library.
In the README.md file I have written how to start the "Picture-Reader", what modules it uses and how to download them.
